feat: add token based auth using JWT - change StockDto to view model - add services to DI container

+++ StockExchange/StockExchange.Server/Services/OrderService.cs
@@ -65,7 +65,7 @@ namespace StockExchange.Server.Services
             }
         }
 
-        private decimal AccountBalance { get; set; }
+        private double AccountBalance { get; set; }
 
         private int GetShareCount(string symbol)
         {
